Move graph to Map components folder.
